[UPDATE] post di

+++ data/src/main/java/com/findapple/findapple/data/dto/request/RegisterRequest.kt
@@ -5,8 +5,8 @@ import com.findapple.findapple.domain.auth.`object`.RegisterObject
 data class RegisterRequest(
     val kakaoId: String,
     val kakaoNickName: String,
-    val longitude: Long,
-    val latitude: Long
+    val longitude: Double,
+    val latitude: Double
 )
 
 fun RegisterObject.toRequest() =
